backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
## python-jose #python-jose[cryptography]
python-dotenv
requests